database transaction refactor
